requirements.txt
allure-pytest==2.9.45
selenium==4.3.0
pytest==7.1.2